Refactored code for chat into a component and reverified functionality.  Made chat have mobile responsiveness on the layout.
